Add files via upload
The overall requirement of this exercise can be summarized:
1. Analyse the adolescent fertility rates data of India over time, identify seasonal patterns,
identify the general trends in the rates and create a time series model
2. Use a time-series model to forecast the revenue in the near future

Data used for the analysis in this assignment accounts for the yearly fertility rates in India
from 1960 – 2019. The annual number of births to women aged 15-19 years per 1,000 women
in that age group. It is also referred to as the age-specific fertility rate for women aged 15-19.
Pregnancy and childbirth are the leading cause of death among adolescent girls aged 15-19.
More than half of the abortions that occur among adolescent girls are unsafe. They also
face higher health risks than adult mothers aged 20-24, including complications from
childbirth contributing to maternal mortality, low baby birth weight, and severe neonatal
conditions.